requirements.txt
fastapi>=0.114,<0.116
uvicorn[standard]>=0.36,<0.37
openai>=1.40
openai-agents>=0.0.3
openai-chatkit>=1.4.0,<2
pydantic>=2.0
